Addresses issues SCRC-239 and SCRC-351
This commit implements multiple changes, with the motivation that
changing datafile extensions from `.sampleCSV` to `.csv` was best
combined with creating a new directory for test fixture data. This
separates files used in testing from their sample data equivalents,
so that the sample data files can be changed independently.

The main changes are:

A new directory containing test input files is created as
`tests/test_data`.

- `tests/conftest.py`: defines `FIXTURE_DIR` as a path to the
fixture data files, and moves from `os.path` to `pathlib.Path`
(cleaning up the code).
- `simple_network_sim/sampleUseOfModel.py`: moves from
`os.path` to `pathlib.Path` for consistency.
- `tests/regression/test_sampleUseOfModel.py`: moves from `os.path`
to `pathlib.Path` for consistenct.

The move to `pathlib.Path` simplifies relative file path
manipulation, but is an opinionated choice and can be reverted.